requirements_librarian.txt
chromadb
sentence-transformers
torch
numpy
langchain-text-splitters
langchain-community
